# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: dawnbx
  version: v1
  description: |
    Self-hosted sandboxes for AI agents. Auth: `Authorization: Bearer dbx_...`
    (API keys) or the `dawnbx_session` cookie (dashboard). Cookie-authed writes
    need `X-Dawnbx: 1`. Errors are `{code, message, hint?}`.
    Terminal: `GET /v1/sandboxes/{id}/terminal` upgrades to a WebSocket
    (subprotocol `dawnbx`); binary frames are the tty, text frames
    `{"cols":N,"rows":N}` resize it. API errors close with 4000+status.
servers:
  - url: /
security:
  - bearer: []
  - cookie: []
paths:
...
  /v1/me:
    get:
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Principal" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/me/password:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old, new]
              properties:
                old: { type: string }
                new: { type: string, minLength: 10, maxLength: 72 }
      responses:
        "204": { description: changed; every session of the user ends }
        default: { $ref: "#/components/responses/Error" }
  /v1/keys:
    get:
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { type: object, required: [keys], properties: { keys: { type: array, items: { $ref: "#/components/schemas/Key" } } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                ttl: { type: string, description: 'Go duration like "720h"; empty = never expires' }
                org: { type: string, description: admins only }
      responses:
        "200":
          description: the token is shown only here
          content:
            application/json:
              schema: { type: object, required: [key, info], properties: { key: { type: string }, info: { $ref: "#/components/schemas/Key" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/keys/{key}:
    parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
    delete:
      responses:
        "204": { description: revoked }
        default: { $ref: "#/components/responses/Error" }
  /v1/audit:
    get:
      responses:
        "200":
          description: latest 200 events
          content:
            application/json:
              schema: { type: object, required: [events], properties: { events: { type: array, items: { $ref: "#/components/schemas/Event" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/orgs:
    get:
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { type: object, required: [orgs], properties: { orgs: { type: array, items: { $ref: "#/components/schemas/Org" } } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string, pattern: "^[a-z0-9][a-z0-9-]{0,39}$" }
                name: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { type: object, required: [id, name], properties: { id: { type: string }, name: { type: string } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/users:
    get:
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { type: object, required: [users], properties: { users: { type: array, items: { $ref: "#/components/schemas/User" } } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, minLength: 10, maxLength: 72 }
                org: { type: string }
                role: { type: string, enum: [member, admin] }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/users/{username}:
    parameters: [{ $ref: "#/components/parameters/username" }]
    delete:
      responses:
        "204": { description: deleted }
        default: { $ref: "#/components/responses/Error" }
  /v1/users/{username}/password:
    parameters: [{ $ref: "#/components/parameters/username" }]
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 10, maxLength: 72 }
      responses:
        "204": { description: changed }
        default: { $ref: "#/components/responses/Error" }
  /v1/nodes:
    get:
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { type: object, required: [nodes], properties: { nodes: { type: array, items: { $ref: "#/components/schemas/Node" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/nodes/join:
    get:
      responses:
        "200":
          description: install command for a new worker (audited)
          content:
            application/json:
              schema: { type: object, required: [command], properties: { command: { type: string } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/nodes/{name}:
    parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
    delete:
      responses:
        "204": { description: removed }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    cookie: { type: apiKey, in: cookie, name: dawnbx_session }
  parameters:
...
    username: { name: username, in: path, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        hint: { type: string }
...
    Principal:
      type: object
      required: [org, admin]
      properties:
        org: { type: string }
        user: { type: string, description: dashboard sessions }
        key_id: { type: string, description: API keys }
        admin: { type: boolean }
    Key:
      type: object
      required: [id, org, name, created, created_by, last_used, expires, revoked]
      properties:
        id: { type: string }
        org: { type: string }
        name: { type: string }
        created: { type: string, format: date-time }
        created_by: { type: string }
        last_used: { type: [string, "null"], format: date-time }
        expires: { type: [string, "null"], format: date-time }
        revoked: { type: [string, "null"], format: date-time }
    Event:
      type: object
      required: [at, org, actor, action, target]
      properties:
        at: { type: string, format: date-time }
        org: { type: string }
        actor: { type: string }
        action: { type: string }
        target: { type: string }
    Org:
      type: object
      required: [id, name, created]
      properties:
        id: { type: string }
        name: { type: string }
        created: { type: string, format: date-time }
    User:
      type: object
      required: [id, org, username, role, created]
      properties:
        id: { type: string }
        org: { type: string }
        username: { type: string }
        role: { type: string, enum: [member, admin] }
        created: { type: string, format: date-time }
    Node:
      type: object
      required: [name, role, ready, since, heartbeat, ip, kubelet, sandboxes]
      properties:
        name: { type: string }
        role: { type: string, enum: [server, worker] }
        ready: { type: boolean }
        since: { type: string, format: date-time, description: last Ready change }
        heartbeat: { type: string, format: date-time }
        ip: { type: string }
        kubelet: { type: string }
        sandboxes: { type: integer }
...